+++ AmanTask/Controllers/ProductController.cs
@@ -8,27 +8,31 @@ using System;
 
 namespace AmanTask.Web.Controllers
 {
-    [Authorize]
     public class ProductController : Controller
     {
         IUnitOfWork _UnitOfWOrk;
         public ProductController(IUnitOfWork UnitOfWOrk)
         {
-            _UnitOfWOrk=UnitOfWOrk;
+            _UnitOfWOrk = UnitOfWOrk;
         }
         // GET: ProductController
         public ActionResult Index()
         {
             try
             {
-                string[] includes = new string[] { "MainCategory", "SubCategory" };
+                if (HttpContext.Session.GetString("_Token") != null)
+                {
+                    string[] includes = new string[] { "MainCategory", "SubCategory" };
 
-                var products = _UnitOfWOrk.GetProductRepo().Get(includes) ;
-                if(products == null)
-                    return NotFound();
+                    var products = _UnitOfWOrk.GetProductRepo().Get(includes);
+                    if (products == null)
+                        return NotFound();
 
-                return View(products);
+                    return View(products);
+                }
 
+
+                return NotFound();
             }
             catch (Exception ex)
             {
@@ -39,29 +43,37 @@ namespace AmanTask.Web.Controllers
         // GET: ProductController/Details/5
         public ActionResult Details(int id)
         {
-            string[] includes = new string[] { "MainCategory", "SubCategory" };
+            if (HttpContext.Session.GetString("_Token") != null)
+            {
+                string[] includes = new string[] { "MainCategory", "SubCategory" };
 
-            if (id == 0)
-                return BadRequest();
-            var productinDb = _UnitOfWOrk.GetProductRepo().GetByID(id,includes);
+                if (id == 0)
+                    return BadRequest();
+                var productinDb = _UnitOfWOrk.GetProductRepo().GetByID(id, includes);
 
-            if (productinDb == null)
-                return NotFound();
-         
+                if (productinDb == null)
+                    return NotFound();
 
-            return View(productinDb);
+
+                return View(productinDb);
+            }
+            return NotFound();
         }
 
         // GET: ProductController/Create
         public ActionResult Create()
         {
-            string[] includes = new string[] { "MainCategory" };
+            if (HttpContext.Session.GetString("_Token") != null)
+            {
+                string[] includes = new string[] { "MainCategory" };
 
-            var mainCat = _UnitOfWOrk.GetMainCategoryRepo().Get();
-            ViewBag.MianCategoryId = new SelectList(mainCat, "Id", "Name");
+                var mainCat = _UnitOfWOrk.GetMainCategoryRepo().Get();
+                ViewBag.MianCategoryId = new SelectList(mainCat, "Id", "Name");
 
-            var SubCat = _UnitOfWOrk.GetSubCategoryRepo().Get(includes);
-            ViewBag.SubCategoryId = new SelectList(SubCat, "Id", "Name");
+                var SubCat = _UnitOfWOrk.GetSubCategoryRepo().Get(includes);
+                ViewBag.SubCategoryId = new SelectList(SubCat, "Id", "Name");
+            }
+
 
             return View();
         }
@@ -73,14 +85,18 @@ namespace AmanTask.Web.Controllers
         {
             try
             {
-                if (ModelState.IsValid)
+                if (HttpContext.Session.GetString("_Token") != null)
                 {
-                    if (product == null)
-                        return BadRequest();
-                    _UnitOfWOrk.GetProductRepo().Add(product);
-                    _UnitOfWOrk.Save();
+                    if (ModelState.IsValid)
+                    {
+                        if (product == null)
+                            return BadRequest();
+                        _UnitOfWOrk.GetProductRepo().Add(product);
+                        _UnitOfWOrk.Save();
+                    }
+                    return RedirectToAction(nameof(Index));
                 }
-                return RedirectToAction(nameof(Index));
+                return NotFound();
             }
             catch
             {
@@ -91,21 +107,25 @@ namespace AmanTask.Web.Controllers
         // GET: ProductController/Edit/5
         public ActionResult Edit(int id)
         {
-            string[] includes = new string[] { "MainCategory" };
+            if (HttpContext.Session.GetString("_Token") != null)
+            {
+                string[] includes = new string[] { "MainCategory" };
 
-            var mainCat = _UnitOfWOrk.GetMainCategoryRepo().Get();
-            ViewBag.MianCategoryId = new SelectList(mainCat, "Id", "Name");
+                var mainCat = _UnitOfWOrk.GetMainCategoryRepo().Get();
+                ViewBag.MianCategoryId = new SelectList(mainCat, "Id", "Name");
 
-            var SubCat = _UnitOfWOrk.GetSubCategoryRepo().Get(includes);
-            ViewBag.SubCategoryId = new SelectList(SubCat, "Id", "Name");
-            if (id == 0)
-                return BadRequest();
-            var product =_UnitOfWOrk.GetProductRepo().GetByID(id);
-            if (product == null)
-                return NotFound();
+                var SubCat = _UnitOfWOrk.GetSubCategoryRepo().Get(includes);
+                ViewBag.SubCategoryId = new SelectList(SubCat, "Id", "Name");
+                if (id == 0)
+                    return BadRequest();
+                var product = _UnitOfWOrk.GetProductRepo().GetByID(id);
+                if (product == null)
+                    return NotFound();
 
 
-            return View(product);
+                return View(product);
+            }
+            return NotFound();
         }
 
         // POST: ProductController/Edit/5
@@ -115,27 +135,32 @@ namespace AmanTask.Web.Controllers
         {
             try
             {
-                string[] includes = new string[] { "MainCategory", "SubCategory" };
-
-                if (ModelState.IsValid)
+                if (HttpContext.Session.GetString("_Token") != null)
                 {
-                    if (id == 0)
-                        return BadRequest();
-                    var productinDb = _UnitOfWOrk.GetProductRepo().GetByID(id, includes);
+                    string[] includes = new string[] { "MainCategory", "SubCategory" };
 
-                    if (productinDb == null || product == null)
-                        return NotFound();
-                    productinDb.Name = product.Name;
-                    productinDb.price = product.price;
-                    productinDb.Qty = product.Qty;
-                    productinDb.SubCategoryId = product.SubCategoryId;
-                    productinDb.MianCategoryId = product.MianCategoryId;
-                    _UnitOfWOrk.GetProductRepo().Update(productinDb);
-                    _UnitOfWOrk.Save();
+                    if (ModelState.IsValid)
+                    {
+                        if (id == 0)
+                            return BadRequest();
+                        var productinDb = _UnitOfWOrk.GetProductRepo().GetByID(id, includes);
+
+                        if (productinDb == null || product == null)
+                            return NotFound();
+                        productinDb.Name = product.Name;
+                        productinDb.price = product.price;
+                        productinDb.Qty = product.Qty;
+                        productinDb.SubCategoryId = product.SubCategoryId;
+                        productinDb.MianCategoryId = product.MianCategoryId;
+                        _UnitOfWOrk.GetProductRepo().Update(productinDb);
+                        _UnitOfWOrk.Save();
+                    }
+
+
+                    return RedirectToAction(nameof(Index));
                 }
-               
+                return NotFound();
 
-                return RedirectToAction(nameof(Index));
             }
             catch
             {
@@ -146,18 +171,23 @@ namespace AmanTask.Web.Controllers
         // GET: ProductController/Delete/5
         public ActionResult Delete(int id)
         {
-            string[] includes = new string[] { "MainCategory", "SubCategory" };
+            if (HttpContext.Session.GetString("_Token") != null)
+            {
+                string[] includes = new string[] { "MainCategory", "SubCategory" };
 
 
-            if (id == 0)
-                return BadRequest();
-            var product = _UnitOfWOrk.GetProductRepo().GetByID(id, includes);
-            if (product == null)
-                return NotFound();
-            _UnitOfWOrk.GetProductRepo().Remove(product);
-            _UnitOfWOrk.Save();
+                if (id == 0)
+                    return BadRequest();
+                var product = _UnitOfWOrk.GetProductRepo().GetByID(id, includes);
+                if (product == null)
+                    return NotFound();
+                _UnitOfWOrk.GetProductRepo().Remove(product);
+                _UnitOfWOrk.Save();
 
-            return View(product);
+                return View(product);
+            }
+            return NotFound();
+
         }
 
         // POST: ProductController/Delete/5
@@ -167,19 +197,24 @@ namespace AmanTask.Web.Controllers
         {
             try
             {
-                string[] includes = new string[] { "MainCategory", "SubCategory" };
+                if (HttpContext.Session.GetString("_Token") != null)
+                {
+                    string[] includes = new string[] { "MainCategory", "SubCategory" };
 
-                if (id == 0)
-                    return BadRequest();
-                var productinDb = _UnitOfWOrk.GetProductRepo().GetByID(id, includes);
+                    if (id == 0)
+                        return BadRequest();
+                    var productinDb = _UnitOfWOrk.GetProductRepo().GetByID(id, includes);
 
-                if (productinDb == null)
-                    return NotFound();
-             
+                    if (productinDb == null)
+                        return NotFound();
 
-                _UnitOfWOrk.GetProductRepo().Remove(productinDb);
-                _UnitOfWOrk.Save();
-                return RedirectToAction(nameof(Index));
+
+                    _UnitOfWOrk.GetProductRepo().Remove(productinDb);
+                    _UnitOfWOrk.Save();
+                    return RedirectToAction(nameof(Index));
+
+                }
+                return NotFound();
 
             }
             catch
